+++ climpred/graphics.py
@@ -1,3 +1,4 @@
+import warnings
 from collections import OrderedDict
 
 import matplotlib as mpl
@@ -6,11 +7,10 @@ import numpy as np
 import xarray as xr
 from xarray.coding.times import infer_calendar_name
 
-from climpred.checks import DimensionError
-from climpred.constants import CLIMPRED_DIMS
-from climpred.utils import get_lead_cftime_shift_args, shift_cftime_index
-
-from .metrics import PROBABILISTIC_METRICS
+from .checks import DimensionError
+from .constants import CLIMPRED_DIMS
+from .metrics import ALL_METRICS, PROBABILISTIC_METRICS
+from .utils import get_lead_cftime_shift_args, get_metric_class, shift_cftime_index
 
 
 def plot_relative_entropy(rel_ent, rel_ent_threshold=None, **kwargs):
@@ -74,32 +74,25 @@ def plot_relative_entropy(rel_ent, rel_ent_threshold=None, **kwargs):
     return ax
 
 
-def plot_bootstrapped_skill_over_leadyear(bootstrapped, plot_persistence=True, ax=None):
+def plot_bootstrapped_skill_over_leadyear(
+    bootstrapped,
+    ax=None,
+    color_initialized="indianred",
+    color_uninitialized="steelblue",
+    color_persistence="gray",
+    color_climatology="tan",
+    capsize=4,
+    fontsize=8,
+    figsize=(10, 4),
+    fmt="--o",
+):
     """
     Plot Ensemble Prediction skill as in Li et al. 2016 Fig.3a-c.
 
     Args:
         bootstrapped (xr.DataArray or xr.Dataset with one variable):
-            from bootstrap_perfect_model or bootstrap_hindcast
+            from PredictionEnsembleEnsemble.bootstrap() or HindcastEnsemble.bootstrap()
 
-            containing:
-        init_skill (xr.Dataset): skill of initialized
-        init_ci (xr.Dataset): confidence levels of init_skill
-        uninit_skill (xr.Dataset): skill of uninitialized
-        uninit_ci (xr.Dataset): confidence levels of uninit_skill
-        p_uninit_over_init (xr.Dataset): p value of the hypothesis that the
-                                         difference of skill between the
-                                         initialized and uninitialized
-                                         simulations is smaller or equal to
-                                         zero based on bootstrapping with
-                                         replacement. Defaults to None.
-        pers_skill (xr.Dataset): skill of persistence
-        pers_ci (xr.Dataset): confidence levels of pers_skill
-        p_pers_over_init (xr.Dataset): p value of the hypothesis that the
-                                       difference of skill between the
-                                       initialized and persistence simulations
-                                       is smaller or equal to zero based on
-                                       bootstrapping with replacement.
         ax (plt.axes): plot on ax. Defaults to None.
 
     Returns:
@@ -126,41 +119,23 @@ def plot_bootstrapped_skill_over_leadyear(bootstrapped, plot_persistence=True, a
             bootstrapped.attrs = attrs
 
     assert isinstance(bootstrapped, xr.DataArray)
+    reference = list(bootstrapped.drop_sel(skill="initialized").coords["skill"].values)
 
     sig = bootstrapped.attrs["confidence_interval_levels"].split("-")
     sig = int(100 * (float(sig[0]) - float(sig[1])))
     pers_sig = sig
 
-    if "metric" in bootstrapped.attrs:
-        if bootstrapped.attrs["metric"] in PROBABILISTIC_METRICS:
-            plot_persistence = False
-
     init_skill = bootstrapped.sel(skill="initialized", results="verify skill")
     init_ci = bootstrapped.sel(
         skill="initialized", results=["low_ci", "high_ci"]
     ).rename({"results": "quantile"})
-    uninit_skill = bootstrapped.sel(skill="uninitialized", results="verify skill")
-    uninit_ci = bootstrapped.sel(
-        skill="uninitialized", results=["low_ci", "high_ci"]
-    ).rename({"results": "quantile"})
-    pers_skill = bootstrapped.sel(skill="persistence", results="verify skill")
-    pers_ci = bootstrapped.sel(
-        skill="persistence", results=["low_ci", "high_ci"]
-    ).rename({"results": "quantile"})
-    p_uninit_over_init = bootstrapped.sel(skill="uninitialized", results="p")
-    p_pers_over_init = bootstrapped.sel(skill="persistence", results="p")
-
-    fontsize = 8
-    c_uninit = "indianred"
-    c_init = "steelblue"
-    c_pers = "gray"
-    capsize = 4
 
     if pers_sig != sig:
         raise NotImplementedError("pers_sig != sig not implemented yet.")
 
     if ax is None:
-        _, ax = plt.subplots(figsize=(10, 4))
+        _, ax = plt.subplots(figsize=figsize)
+    # plot init
     ax.errorbar(
         init_skill.lead,
         init_skill,
@@ -168,79 +143,52 @@ def plot_bootstrapped_skill_over_leadyear(bootstrapped, plot_persistence=True, a
             init_skill - init_ci.isel(quantile=0),
             init_ci.isel(quantile=1) - init_skill,
         ],
-        fmt="--o",
+        fmt=fmt,
         capsize=capsize,
-        c=c_uninit,
-        label=(" ").join(["initialized with", str(sig) + "%", "confidence interval"]),
+        c=color_initialized,
+        label="initialized",
     )
-    # uninit
-    if p_uninit_over_init is not None:
-        # add p values
+    # plot references
+    for r in reference:
+        r_skill = bootstrapped.sel(skill=r, results="verify skill")
+        if (r_skill == np.nan).all():
+            warnings.warn(f"Found only NaNs in {r} verify skill and skipped.")
+            continue
+        p_r_over_init = bootstrapped.sel(skill=r, results="p")
+        r_ci = bootstrapped.sel(skill=r, results=["low_ci", "high_ci"]).rename(
+            {"results": "quantile"}
+        )
+        c = eval(f"color_{r}")
+        # add p values over all reference skills
         for t in init_skill.lead.values:
             ax.text(
-                init_skill.lead.sel(lead=t),
-                init_ci.isel(quantile=1).sel(lead=t).values,
-                "%.2f" % float(p_uninit_over_init.sel(lead=t).values),
+                r_skill.lead.sel(lead=t),
+                r_ci.isel(quantile=0).sel(lead=t).values,
+                "%.2f" % float(p_r_over_init.sel(lead=t).values),
                 horizontalalignment="center",
                 verticalalignment="bottom",
                 fontsize=fontsize,
-                color=c_uninit,
+                color=c,
             )
-        uninit_skill = uninit_skill.dropna("lead").squeeze()
-        uninit_ci = uninit_ci.dropna("lead").squeeze()
-        if "lead" not in uninit_skill.dims:
-            yerr = [
-                [uninit_skill - uninit_ci.isel(quantile=0)],
-                [uninit_ci.isel(quantile=1) - uninit_skill],
-            ]
-            ax.axhline(y=uninit_skill, c="steelblue", ls=":")
-            x = 0
-        else:
-            yerr = [
-                uninit_skill - uninit_ci.isel(quantile=0),
-                uninit_ci.isel(quantile=1) - uninit_skill,
-            ]
-            x = uninit_skill.lead
+        yerr = [
+            r_skill - r_ci.isel(quantile=0),
+            r_ci.isel(quantile=1) - r_skill,
+        ]
+        x = r_skill.lead
         ax.errorbar(
             x,
-            uninit_skill,
+            r_skill,
             yerr=yerr,
-            fmt="--o",
+            fmt=fmt,
             capsize=capsize,
-            c=c_init,
-            label=(" ").join(
-                ["uninitialized with", str(sig) + "%", "confidence interval"]
-            ),
+            c=c,
+            label=r,
         )
-    # persistence
-    if plot_persistence:
-        if pers_skill is not None and pers_ci is not None:
-            ax.errorbar(
-                pers_skill.lead,
-                pers_skill,
-                yerr=[
-                    pers_skill - pers_ci.isel(quantile=0),
-                    pers_ci.isel(quantile=1) - pers_skill,
-                ],
-                fmt="--o",
-                capsize=capsize,
-                c=c_pers,
-                label=f"persistence with {pers_sig}% confidence interval",
-            )
-        for t in pers_skill.lead.values:
-            ax.text(
-                pers_skill.lead.sel(lead=t),
-                pers_ci.isel(quantile=0).sel(lead=t).values,
-                "%.2f" % float(p_pers_over_init.sel(lead=t).values),
-                horizontalalignment="center",
-                verticalalignment="bottom",
-                fontsize=fontsize,
-                color=c_pers,
-            )
 
-    ax.xaxis.set_ticks(np.arange(init_skill.lead.size + 1))
-    ax.legend(frameon=False)
-    ax.set_xlabel("Lead time [years]")
+    ax.xaxis.set_ticks(bootstrapped.lead.values)
+    ax.legend(frameon=False, title=f"skill with {sig}% confidence interval:")
+    ax.set_xlabel(f"Lead time [{bootstrapped.lead.attrs['units']}]")
+    ax.set_ylabel(get_metric_class(bootstrapped.attrs["metric"], ALL_METRICS).long_name)
     return ax