+++ climpred/tests/test_probabilistic.py
@@ -26,95 +26,116 @@ probabilistic_metrics_requiring_more_than_member_dim = [
     "reliability",
 ]
 
+references = [
+    [],
+    "uninitialized",
+    "persistence",
+    "climatology",
+    ["climatology", "uninitialized", "persistence"],
+]
+references_ids = [
+    "empty list",
+    "uninitialized",
+    "persistence",
+    "climatology",
+    "climatology, uninitialized, persistence",
+]
+
 xr.set_options(display_style="text")
 
 
-@pytest.mark.parametrize("comparison", PROBABILISTIC_PM_COMPARISONS)
-@pytest.mark.parametrize("metric", PROBABILISTIC_METRICS)
-def test_compute_perfect_model_da1d_not_nan_probabilistic(
-    PM_da_initialized_1d, PM_da_control_1d, metric, comparison
-):
-    """
-    Checks that there are no NaNs on perfect model probabilistic metrics of 1D
-    time series.
-    """
-    metric_kwargs = {"comparison": comparison, "metric": metric, "dim": "member"}
-    if "threshold" in metric:
-        metric_kwargs["threshold"] = 10.5
-    if metric == "brier_score":
-
-        def func(x):
-            return x > 0
-
-        metric_kwargs["logical"] = func
-
-    actual = compute_perfect_model(PM_da_initialized_1d, PM_da_control_1d)
-    actual = actual.isnull().any()
-    assert not actual
-
-
+@pytest.mark.parametrize("reference", references, ids=references_ids)
 @pytest.mark.parametrize("metric", PROBABILISTIC_METRICS)
 @pytest.mark.parametrize("comparison", PROBABILISTIC_HINDCAST_COMPARISONS)
-def test_compute_hindcast_probabilistic(hindcast_recon_1d_ym, metric, comparison):
+def test_HindcastEnsemble_verify_bootstrap_probabilistic(
+    hindcast_hist_obs_1d, metric, comparison, reference
+):
     """
-    Checks that compute hindcast works without breaking.
+    Checks that HindcastEnsemble.verify() and HindcastEnsemble.bootstrap() works
+    without breaking for all probabilistic metrics.
     """
-    category_edges = np.array([0, 0.5, 1])
+    he = hindcast_hist_obs_1d.isel(lead=[0, 1, 2])
+
+    category_edges = np.array([-0.5, 0, 0.5])
     if metric in probabilistic_metrics_requiring_logical:
 
         def f(x):
-            return x > 0.5
+            return x > 0
 
-        metric_kwargs = {"logical": f}
+        kwargs = {"logical": f}
     elif metric == "threshold_brier_score":
-        metric_kwargs = {"threshold": 0.5}
+        kwargs = {"threshold": 0}
     elif metric == "contingency":
-        metric_kwargs = {
+        kwargs = {
             "forecast_category_edges": category_edges,
             "observation_category_edges": category_edges,
             "score": "accuracy",
         }
     elif metric == "rps":
-        metric_kwargs = {"category_edges": category_edges}
+        kwargs = {"category_edges": category_edges}
     else:
-        metric_kwargs = {}
+        kwargs = {}
     dim = (
         ["member", "init"]
         if metric in probabilistic_metrics_requiring_more_than_member_dim
         else "member"
     )
-    res = hindcast_recon_1d_ym.verify(
-        alignment="same_verif",
-        comparison=comparison,
-        metric=metric,
-        dim=dim,
-        **metric_kwargs,
-    )["SST"]
-    assert not res.isnull().all()
+    # verify()
+    kwargs.update(
+        {
+            "comparison": comparison,
+            "metric": metric,
+            "dim": dim,
+            "reference": reference,
+            "alignment": "same_verifs",
+        }
+    )
+    actual_verify = he.verify(**kwargs)["SST"]
+    not actual_verify.isnull().all()
+
+    # bootstrap()
+    actual = he.bootstrap(iterations=3, **kwargs)["SST"]
+    assert "dayofyear" not in actual.coords
+
+    if isinstance(reference, str):
+        reference = [reference]
+    if len(reference) == 0:
+        assert not actual.sel(results="verify skill").isnull().all()
+    else:
+        assert (
+            not actual.sel(skill="initialized", results="verify skill").isnull().all()
+        )
+        for skill in reference:
+            actual_skill = actual.sel(skill=skill, results="verify skill")
+            if metric == "crpss_es" and skill in ["climatology", "persistence"]:
+                pass
+            else:
+                assert not actual_skill.isnull().all()
 
 
+@pytest.mark.parametrize("reference", references, ids=references_ids)
 @pytest.mark.parametrize("comparison", PROBABILISTIC_PM_COMPARISONS)
 @pytest.mark.parametrize("metric", PROBABILISTIC_METRICS)
-def test_bootstrap_perfect_model_da1d_not_nan_probabilistic(
-    PM_da_initialized_1d, PM_da_control_1d, metric, comparison
+def test_PerfectModelEnsemble_verify_bootstrap_not_nan_probabilistic(
+    perfectModelEnsemble_initialized_control, metric, comparison, reference
 ):
     """
-    Checks that there are no NaNs on perfect model probabilistic metrics of 1D
-    time series.
+    Checks that PerfectModelEnsemble.verify() and PerfectModelEnsemble.bootstrap() works without breaking for all probabilistic metrics.
     """
+    pm = perfectModelEnsemble_initialized_control.isel(lead=[0, 1, 2, 3])
     kwargs = {
         "comparison": comparison,
         "metric": metric,
     }
-    category_edges = np.array([0, 0.5, 1])
+    category_edges = np.array([9.5, 10.0, 10.5])
     if metric in probabilistic_metrics_requiring_logical:
 
         def f(x):
-            return x > 0.5
+            return x > 10
 
         kwargs["logical"] = f
     elif metric == "threshold_brier_score":
-        kwargs["threshold"] = 0.5
+        kwargs["threshold"] = 10.0
     elif metric == "contingency":
         kwargs["forecast_category_edges"] = category_edges
         kwargs["observation_category_edges"] = category_edges
@@ -128,63 +149,26 @@ def test_bootstrap_perfect_model_da1d_not_nan_probabilistic(
     )
     kwargs["dim"] = dim
 
-    assert (
-        not compute_perfect_model(PM_da_initialized_1d, PM_da_control_1d, **kwargs)
-        .isnull()
-        .all()
-    )
+    # verify()
+    actual_verify = pm.verify(**kwargs)
+    assert not actual_verify.tos.isnull().all()
 
+    # bootstrap
     kwargs["iterations"] = ITERATIONS
     kwargs["resample_dim"] = "member"
-    actual = bootstrap_perfect_model(PM_da_initialized_1d, PM_da_control_1d, **kwargs)
-    for skill in ["initialized", "uninitialized"]:
-        actualk = actual.sel(skill=skill, results="verify skill")
-        actualk = actualk.isnull().all()
-        assert not actualk
-
-
-@pytest.mark.slow
-@pytest.mark.parametrize("comparison", PROBABILISTIC_HINDCAST_COMPARISONS)
-@pytest.mark.parametrize("metric", PROBABILISTIC_METRICS)
-def test_bootstrap_hindcast_da1d_not_nan_probabilistic(
-    hind_da_initialized_1d,
-    hist_da_uninitialized_1d,
-    observations_da_1d,
-    metric,
-    comparison,
-):
-    """
-    Checks that there are no NaNs on hindcast probabilistic metrics of 1D
-    time series.
-    """
-    metric_kwargs = {
-        "comparison": comparison,
-        "metric": metric,
-        "dim": "member",
-        "iterations": ITERATIONS,
-        "alignment": "same_verif",
-    }
-    if "threshold" in metric:
-        metric_kwargs["threshold"] = 0.5
-    if metric == "brier_score":
-
-        def func(x):
-            return x > 0
-
-        metric_kwargs["logical"] = func
-
-    actual = bootstrap_hindcast(
-        hind_da_initialized_1d,
-        hist_da_uninitialized_1d,
-        observations_da_1d,
-        resample_dim="member",
-    )
-    for skill in ["initialized", "uninitialized"]:
-        actualk = actual.sel(skill=skill, results="verify skill")
-        if "init" in actualk.coords:
-            actualk = actualk.mean("init")
-        actualk = actualk.isnull().any()
-        assert not actualk
+    kwargs["reference"] = reference
+    actual = pm.bootstrap(**kwargs).tos
+    if isinstance(reference, str):
+        reference = [reference]
+    if len(reference) == 0:
+        assert not actual.sel(results="verify skill").isnull().all()
+    else:
+        for skill in reference:
+            actual_skill = actual.sel(skill=skill, results="verify skill")
+            if metric == "crpss_es" and skill in ["climatology", "persistence"]:
+                pass
+            else:
+                assert not actual_skill.isnull().all()
 
 
 def test_compute_perfect_model_da1d_not_nan_crpss_quadratic(
@@ -244,7 +228,7 @@ def test_compute_hindcast_da1d_not_nan_crpss_quadratic(
     """
     actual = (
         compute_hindcast(
-            hind_da_initialized_1d,
+            hind_da_initialized_1d.isel(lead=[0, 1, 2]),
             observations_da_1d,
             comparison="m2o",
             metric="crpss",
@@ -318,6 +302,20 @@ def test_compute_hindcast_probabilistic_metric_e2o_fails(
             metric=metric,
             dim="member",
         )
+
+
+def test_HindcastEnsemble_rps_terciles(hindcast_hist_obs_1d):
+    actual = hindcast_hist_obs_1d.verify(
+        metric="rps",
+        comparison="m2o",
+        dim=["member", "init"],
+        alignment="same_verifs",
+        category_edges=np.array([-0.5, 0.0, 0.5, 1]),
+        reference="climatology",
+    )  # todo really use terciles
+    assert actual.notnull().all()
+    rpss = 1 - actual.sel(skill="initialized") / actual.sel(skill="climatology")
+    assert rpss.isel(lead=0) > 0
 
 
 def test_hindcast_verify_brier_logical(hindcast_recon_1d_ym):